Switched everything over to Python 3.5.   This makes me sad.
Most of the changes were making the code worse -- but unfortunately if
we're going to promote use of these scripts, we'll need to support
people on the "classic" image, who don't have access to 3.6.   Requiring
them to update their system image (potentially breaking production
sites) in order to use the scripts means that they just won't use them.